Fix nul-termination and buffer overrun error.
read_guest_null_terminated_string() could read past the end of the
allocated buffer if either no nul character was found in the first
1024 bytes, or if a memory read failed as the string was not
terminated.

Fix the issue by using the std::string constructor that takes a
size.

At the same time, avoid reading the string twice for no reason.

In addition, panda_virtual_memory_read() returns either 0, -1, or
a return value from panda_physical_memory_rw() which is either
MEMTX_OK or MEMTX_ERROR. MEMTX_OK is conveniently 0, but
MEMTX_ERROR is not -1 (nor is MEMTX_DECODE_ERROR) so checking for
failure needs to check that's it's not 0 (and also not MEMTX_OK).
Given they're the same, we just check 0.